♻️  core: Split idl::parse module a bit
